requirements.txt
torch
torchaudio
openai-whisper
fastapi
uvicorn
